+++ source-code/clj/server_plugins/item_editor/api.clj
@@ -14,7 +14,6 @@
 
 ; server-plugins.item-editor.engine
 (def value-path engine/value-path)
-(def form-id    engine/form-id)
 
 ; server-plugins.item-editor.resolvers
 (def get-item-suggestions resolvers/get-item-suggestions)